refactor : 패키지 리네임 (step3 -> racingcar)
