+++ src/main/kotlin/com/nextstep/jngcii/racingcar/domain/Dice.kt
@@ -1,0 +1,5 @@
+package com.nextstep.jngcii.racingcar.domain
+
+interface Dice {
+    fun run(): Boolean
+}
