--- src/main/kotlin/com/nextstep/jngcii/step3/domain/Dice.kt
@@ -1,5 +1,0 @@
-package com.nextstep.jngcii.step3.domain
-
-interface Dice {
-    fun run(): Boolean
-}
